+++ lib/graph/point.dart
@@ -12,8 +12,6 @@ class GraphPoint {
 
   @override
   bool operator ==(Object other) {
-    if (identical(this, other)) return true;
-  
     return other is GraphPoint &&
       other.id == id &&
       other.x == x &&