+++ lib/grid/state.dart
@@ -1,4 +1,5 @@
 import 'dart:collection';
+import 'package:collection/equality.dart';
 import 'package:gridfind/gridfind.dart';
 
 const dirs4 = [(1, 0), (0, 1), (-1, 0), (0, -1)];
@@ -41,5 +42,22 @@ abstract class GridState {
         p.x >= width ||
         p.y >= height ||
         p.get(grid) == GridNode.obstacle;
+  }
+
+  @override
+  bool operator ==(Object other) {
+    if (other is GridState) {
+      final mapEquality = const MapEquality();
+      final listEquality = const ListEquality();
+      final deepCollectionEquality = const DeepCollectionEquality();
+
+      return other.start == start &&
+          other.target == target &&
+          deepCollectionEquality.equals(other.grid, grid) &&
+          mapEquality.equals(other.parents, parents) &&
+          listEquality.equals(other.dirs, dirs) &&
+          other.status == status;
+    }
+    return false;
   }
 }